Move investment votes partial to a component
This way changing it will be easier.
