--- app/helpers/accessibility_helper.rb
@@ -1,5 +1,0 @@
-module AccessibilityHelper
-  def css_for_aria_hidden(reason)
-    reason.present? ? "true" : ""
-  end
-end
